agent menu refactor
